reorganized tests under one folder. added robots.txt. added sitemap.xml. added static files under public folder
