+++ chathaven/README.md
@@ -48,12 +48,6 @@ npx tree-cli --ignore node_modules --ignore build -l 999
 
 ### `npm run storybook`
 
-Runs the storybook server
-```
-npx start-storybook -p 6006
-``` 
-ref: https://storybook.js.org/docs/react/get-started/start-storybook
-
 ### `npm run tailwindcss-build-watch`
 Compiles the tailwindcss styles and watches for changes
 ```
@@ -67,34 +61,10 @@ Runs the app in production mode on localhost
 npm run build && serve -s build
 ```
 ref: https://www.npmjs.com/package/serve
+
+### /chathaven tree
+```angular2html
+
+```
 ___ 
 
-## Learn More
-
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
-
-### Code Splitting
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/code-splitting](https://facebook.github.io/create-react-app/docs/code-splitting)
-
-### Analyzing the Bundle Size
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size)
-
-### Making a Progressive Web App
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app](https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app)
-
-### Advanced Configuration
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/advanced-configuration](https://facebook.github.io/create-react-app/docs/advanced-configuration)
-
-### Deployment
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/deployment](https://facebook.github.io/create-react-app/docs/deployment)
-
-### `npm run build` fails to minify
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
