+++ chathaven/public/sitemap.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+
+    <url>
+        <loc>https://project2048-soen-341-project-w25.vercel.app/</loc>
+        <lastmod>2025-04-12</lastmod>
+        <changefreq>weekly</changefreq>
+        <priority>1.0</priority>
+    </url>
+
+    <url>
+        <loc>https://project2048-soen-341-project-w25.vercel.app/about</loc>
+        <lastmod>2025-04-12</lastmod>
+        <changefreq>monthly</changefreq>
+        <priority>0.8</priority>
+    </url>
+
+    <url>
+        <loc>https://project2048-soen-341-project-w25.vercel.app/contact</loc>
+        <lastmod>2025-04-12</lastmod>
+        <changefreq>monthly</changefreq>
+        <priority>0.7</priority>
+    </url>
+
+    <url>
+        <loc>https://project2048-soen-341-project-w25.vercel.app/game</loc>
+        <lastmod>2025-04-12</lastmod>
+        <changefreq>daily</changefreq>
+        <priority>0.9</priority>
+    </url>
+
+</urlset>
