+++ chathaven/src/Pages/AdminView.jsx
@@ -1,4 +1,4 @@
-import "./index.css";
+import "../styles/index.css";
 import { useEffect } from "react";
 
 export const AdminView = () => {
