+++ chathaven/src/Pages/MemberView.jsx
@@ -1,4 +1,4 @@
-import "./index.css";
+import "../styles/index.css";
 
 export const MemberView = () => {
   return (
