+++ chathaven/src/Pages/SocialButton.jsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import "./index.css";
+import "../styles/index.css";
 function SocialButton({ src, alt }) {
     return (
         <button
