+++ chathaven/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './root_index.css';
+import './styles/root_index.css';
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
